+++ RELEASE-NOTES.md
@@ -1,6 +1,22 @@
+kount-access-dotnet-sdk 4.0.0
+===========================
+**10/08/2018**
+
+### New features
+* default API version is changed from 2.1 to 4.0
+* added get Kount Access data for session based on what was requested in the info flag
+
+### Improvements
+* updated examples
+* added more unit tests
+* changed Accept HTTP header value from text/json to application/json
+
+### Bugfixes
+- fixed problem with Content Type HTTP header on POST requests
+
 kount-access-dotnet-sdk 1.0.0
 ===========================
-**07/14/2017**
+**14/07/2017**
 
 ### New features
 * added first version of *Kount Access .Net SDK*